requirements.txt
openai
langchain
langchain-openai
langchain-community
langchain-huggingface
langchain-chroma
langchain-google-genai
chromadb
pypdf
pymupdf
tiktoken
colorama
python-dotenv
pydantic
sentence-transformers
tqdm
